src/polygon-web3-client/requirements.txt
web3==5.29.1
PyYAML==5.4.1